Split Named and NamedList and expose them
