+++ core/rpc/mu-rpc.cabal
@@ -1,5 +1,5 @@
 name:               mu-rpc
-version:            0.5.0.1
+version:            0.5.0.2
 synopsis:           Protocol-independent declaration of services and servers.
 description:
   Protocol-independent declaration of services and servers for mu-haskell.
@@ -22,6 +22,7 @@ source-repository head
 
 library
   exposed-modules:
+    Mu.Named
     Mu.Rpc
     Mu.Rpc.Annotations
     Mu.Rpc.Examples
